+++ packages/common/core-utils/src/internal.ts
@@ -3,15 +3,9 @@
  * Licensed under the MIT License.
  */
 
-// The "internal" exports are a superset of the standard ones. So, we want to export everything from the standard barrel file.
-// eslint-disable-next-line no-restricted-syntax, @typescript-eslint/no-restricted-imports
-export * from "./index.js";
+// eslint-disable-next-line no-restricted-syntax -- This internal entrypoint includes all exports shared with the external entrypoint.
+export * from "./main.js";
 
-import { assert as publicAssert } from "./assert.js";
-
-/**
- * Asserts the specified condition.
- *
- * @internal
- */
-export const assert: typeof publicAssert = publicAssert;
+// Export internal variant of `assert`.
+// TODO: when the deprecated external variant is removed, this special barrel file should be removed.
+export { assertInternal as assert } from "./assert.js";
