+++ packages/common/core-utils/src/main.ts
@@ -1,0 +1,40 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+// Note: `assert` is omitted here intentionally.
+// The internal variant is explicitly exported by `internal.ts`, and the deprecated external variant is exported by `index.ts`.
+export {
+	appendDebugMessage,
+	fail,
+	debugAssert,
+	configureDebugAsserts,
+	nonProductionConditionalsIncluded,
+	emulateProductionBuild,
+	onAssertionFailure,
+} from "./assert.js";
+export { compareArrays } from "./compare.js";
+export { delay } from "./delay.js";
+export type { IComparer, IHeapNode } from "./heap.js";
+export { Heap, NumberComparer } from "./heap.js";
+export { Lazy, LazyPromise } from "./lazy.js";
+export {
+	DoublyLinkedList,
+	type ListNode,
+	type ListNodeRange,
+	iterateListValuesWhile,
+	walkList,
+} from "./list.js";
+export type { PromiseCacheExpiry, PromiseCacheOptions } from "./promiseCache.js";
+export { PromiseCache } from "./promiseCache.js";
+export { Deferred } from "./promises.js";
+export { shallowCloneObject } from "./shallowClone.js";
+export type { IPromiseTimer, IPromiseTimerResult, ITimer } from "./timer.js";
+export { PromiseTimer, setLongTimeout, Timer } from "./timer.js";
+export { unreachableCase } from "./unreachable.js";
+export { isObject, isPromiseLike } from "./typesGuards.js";
+export { oob } from "./oob.js";
+export { transformMapValues } from "./map.js";
+export { clamp } from "./math.js";
+export { validateAllOrNone, type AllOrNoneResult } from "./allOrNone.js";
