+++ packages/common/core-utils/src/math.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { assert } from "./assert.js";
+import { assertInternal as assert } from "./assert.js";
 
 /**
  * Clamps `value` to the inclusive range [`min`, `max`].
